Splited statistics tabs to components
Added component for each tab
