+++ app/components/statistics/statisicsShifts/statisticsShifts.component.ts
@@ -1,0 +1,14 @@
+import { Component, Input } from '@angular/core';
+
+declare var d3: any;
+
+@Component({
+    selector: 'statisticsShifts',
+    templateUrl: './statisticsShifts.html',
+    providers: [],
+    styleUrls: ['./statisticsShifts.css']
+})
+
+export class StatisticsShiftsComponent {
+    @Input() workers: Array<any>;
+}
