+++ app/components/statistics/statisicsShifts/statisticsShifts.html
@@ -1,0 +1,1 @@
+<h4>משמרות</h4>
