+++ app/components/statistics/statisticsAges/statisticsAges.html
@@ -1,0 +1,4 @@
+<h4>גילאים</h4>
+<h5 *ngIf="workers">חתך גילאי {{ workers.length }} עובדים</h5>
+<span class="agesInfo" *ngIf="averageAge">ממוצע גילאי עובדים בעסק: <b>{{ averageAge }}</b></span>
+<canvas id="workersAgesChart" width="400" height="400"></canvas>
