+++ app/components/statistics/statisticsConstraints/statisticsConstraints.html
@@ -1,0 +1,1 @@
+<h4>אילוצים</h4>
