+++ app/components/statistics/statisticsSalaries/statisticsSalaries.html
@@ -1,0 +1,3 @@
+<h4>גרף שכר</h4>
+<span *ngIf="averageSalary" class="row justify-content-center">ממוצע שכר לשעה עבור עובדים בעסק: <b>{{ averageSalary }}</b></span>
+<svg id="workersSalariesChart" width="600" height="400"></svg>
